#  pkg/labels: print all leaf CIDRs, not just the last one. 