--- a/pkg/labels/cidr/doc.go
+++ b/pkg/labels/cidr/doc.go
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-// Package cidr provides helper methods for generating labels for CIDRs which
-// are partially derived from node state.
-package cidr