@squeed 2023-09-25
Confirmed that this works as expected:

```
root@kind-control-plane:/home/cilium# cilium identity list
ID         LABELS
1          cidr:10.244.0.80/32
           cidr:172.19.0.2/32
           cidr:fc00:c111::2/128
           reserved:host
           reserved:kube-apiserver
```