@squeed 2023-10-19
Rebased, refactored to use a real data structure. This improves performance over the existing, incorrect method as well as my previous naive attempt:

```
     │ GetPrintableModelOrig │      GetPrintableModelLeafCIDR      │       GetPrintableModelNaive        │
     │        sec/op         │   sec/op     vs base                │   sec/op     vs base                │
*-12            167.0µ ± 17%   108.8µ ± 3%  -34.84% (p=0.000 n=10)   285.4µ ± 5%  +70.93% (p=0.000 n=10)

     │ GetPrintableModelOrig │     GetPrintableModelLeafCIDR      │       GetPrintableModelNaive       │
     │       allocs/op       │ allocs/op   vs base                │ allocs/op   vs base                │
*-12             1440.0 ± 0%   557.0 ± 0%  -61.32% (p=0.000 n=10)   189.0 ± 0%  -86.88% (p=0.000 n=10)

```

This reduces time from 167µs to 108µs. Fewer allocs, too.