@squeed 2023-10-20
Code golf: got it even faster :-)

```
                     │   orig.txt    │               new.txt               │
                     │    sec/op     │   sec/op     vs base                │
GetPrintableModel-12   166.98µ ± 17%   88.62µ ± 3%  -46.93% (p=0.000 n=10)

                     │  orig.txt   │              new.txt               │
                     │  allocs/op  │ allocs/op   vs base                │
GetPrintableModel-12   1440.0 ± 0%   401.0 ± 0%  -72.15% (p=0.000 n=10)
```